Refactor playground shell into registry, composable, and small chrome components
